P4feat(batch): Implement Hashtag, Visitor, Dept stats batches and Idea/Popup APIs

+++ backend/src/main/java/com/kz/magazine/dto/dashboard/VisitorTrendDto.java
@@ -13,6 +13,6 @@ import java.time.LocalDate;
 @AllArgsConstructor
 public class VisitorTrendDto {
     private LocalDate date;
-    private Integer pageViews;
-    private Integer visitorCount; // Optional if we track unique users per day
+    private Long pageViews;
+    private Long visitorCount; // Optional if we track unique users per day
 }
